fix forward

+++ prob.py
@@ -286,8 +286,7 @@ class ProbabilisticUnet(nn.Module):
         # Here we use the posterior sample sampled above
         self.reconstruction = self.reconstruct(use_posterior_mean=reconstruct_posterior_mean, calculate_posterior=False,
                                                z_posterior=z_posterior)
-
-        reconstruction_loss = criterion(input=self.reconstruction, target=segm)
+        reconstruction_loss = criterion(input=self.reconstruction, target=segm.float())
         self.reconstruction_loss = torch.sum(reconstruction_loss)
         self.mean_reconstruction_loss = torch.mean(reconstruction_loss)
